+++ CHANGELOG.md
@@ -8,3 +8,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+  - Add support for Cloud SQL monitoring:
+    - CPU usage.
+    - Storage usage.
+    - Memory usage.
+    - Storage high growth.